code compiling, needs tests, lots of tests!
