+++ Esp.Net/Reactive/IEventObservationRegistrar.cs
@@ -1,0 +1,8 @@
+﻿namespace Esp.Net.Reactive
+{
+    public interface IEventObservationRegistrar
+    {
+        void IncrementRegistration<TEvent>();
+        void DecrementRegistration<TEvent>();
+    }
+}
